bnxt: Re-structure the bnxt_ntuple_filter structure.
With the new bnxt_l2_filter structure, we can now re-structure the
bnxt_ntuple_filter structure to point to the bnxt_l2_filter structure.
We eliminate the L2 ether address info from the ntuple filter structure
as we can get the information from the L2 filter structure.

We hash all the ntuple filters so they can be searched quickly.  The
current code uses the Toeplitz hash and it only works for incoming
packet whose Toeplitz hash is calculated by the NIC.  We switch to
call flow_hash_from_keys() so that we can hash all ntuple filters
including the ones defined by the user.

Signed-off-by: Michael Chan <michael.chan@broadcom.com>
Reviewed-by: Vasundhara Volam <vasundhara-v.volam@broadcom.com>
Signed-off-by: Pavan Chebbi <pavan.chebbi@broadcom.com>