bnxt: Free ntuple filters on a VF when the VF VNIC is freed.
In bnxt_cfg_ntp_filters(), we add logic to go through the L2 filters
belonging to VFs and delete them when the VNIC has been freed.

When we get an async message informing us that the VF's VNIC has been
freed, we'll schedule bnxt_sp_task() to do the deletion.

Signed-off-by: Michael Chan <michael.chan@broadcom.com>
Reviewed-by: Pavan Chebbi <pavan.chebbi@broadcom.com>
Reviewed-by: Vasundhara Volam <vasundhara-v.volam@broadcom.com>
Signed-off-by: Pavan Chebbi <pavan.chebbi@broadcom.com>