bnxt: Add ntuple matching flags to the bnxt_ntuple_filter structure.
Up until now, we only match all 5 tuples for ntuple filters.  In
preparation to support user defined ntuple filters where some of the
tuples can be wildcards, we add the ntuple_flags to the
bnxt_ntuple_filter struct to specify which tuple fields are to be
matched.  The matching tuple fields will then be passed to the
firmware in bnxt_hwrm_cfa_ntuple_filter_alloc() to create the
proper filter.

This patch only adds the new flags and the associated logic.  All filters
are still created to be 5-tuple filters with all matching fields set.

Signed-off-by: Michael Chan <michael.chan@broadcom.com>
Reviewed-by: Vasundhara Volam <vasundhara-v.volam@broadcom.com>
Signed-off-by: Pavan Chebbi <pavan.chebbi@broadcom.com>