net: Proxy net_mp_{open,close}_rxq for mapped queues
When a process in a container wants to setup a memory provider, it will
use the virtual netdev and an rx queue that is a peer queue of a queue
pair. Calls to net_mp_{open,close}_rxq to set up a memory provider and
restart the queue will be transparently proxied to the peer queue in a
real netdev.

Signed-off-by: David Wei <dw@davidwei.uk>
Co-developed-by: Daniel Borkmann <daniel@iogearbox.net>
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>