netkit: Implement rtnl_link_ops->alloc and ndo_queue_create
Implement rtnl_link_ops->alloc that allows the number of rx queues to be
set when netkit is created. By default, netkit has only a single rxq
(and single txq). The number of queues is deliberately not allowed to be
changed via ethtool -L and is fixed for the lifetime of a netkit
instance.

For netkit device creation, numrxqueues with larger than one rxq can be
specified. These rxqs can then be bound to real rxqs in physical
netdevs:

  ip link add numrxqueues 2 type netkit

The limit of numrxqueues for netkit is currently set to 256, which
allows binding multiple real rxqs from physical netdevs.

The implementation of ndo_queue_create() adds a new rxq during the bind
queue operation. We allow to create queues either in single device mode
or for the case of dual device mode for the netkit peer device which
gets placed into the target network namespace. For dual device mode the
bind against the primary device does not make sense for the targeted use
cases, and therefore gets rejected.

Signed-off-by: David Wei <dw@davidwei.uk>
Co-developed-by: Daniel Borkmann <daniel@iogearbox.net>
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>